+++ arches/app/utils/permission_backend.py
@@ -70,10 +70,10 @@ class PermissionFramework(metaclass=ABCMeta):
     ): ...
 
     @abstractmethod
-    def get_groups_for_object(self, perm, obj): ...
+    def get_groups_with_permission_for_object(self, perm, obj): ...
 
     @abstractmethod
-    def get_users_for_object(self, perm, obj): ...
+    def get_users_with_permission_for_object(self, perm, obj): ...
 
     @abstractmethod
     def check_resource_instance_permissions(self, user, resourceid, permission): ...
@@ -162,11 +162,11 @@ def _get_permission_framework():
             )
             _PERMISSION_FRAMEWORK = PermissionFramework()
         else:
-            from arches.app.permissions.arches_standard import (
-                ArchesStandardPermissionFramework,
+            from arches.app.permissions.arches_default_allow import (
+                ArchesDefaultAllowPermissionFramework,
             )
 
-            _PERMISSION_FRAMEWORK = ArchesStandardPermissionFramework()
+            _PERMISSION_FRAMEWORK = ArchesDefaultAllowPermissionFramework()
     return _PERMISSION_FRAMEWORK
 
 
@@ -200,12 +200,12 @@ def get_restricted_instances(user, search_engine=None, allresources=False):
     )
 
 
-def get_groups_for_object(perm, obj):
-    return _get_permission_framework().get_groups_for_object(perm, obj)
+def get_groups_with_permission_for_object(perm, obj):
+    return _get_permission_framework().get_groups_with_permission_for_object(perm, obj)
 
 
-def get_users_for_object(perm, obj):
-    return _get_permission_framework().get_users_for_object(perm, obj)
+def get_users_with_permission_for_object(perm, obj):
+    return _get_permission_framework().get_users_with_permission_for_object(perm, obj)
 
 
 def check_resource_instance_permissions(user, resourceid, permission):
